Added the new version on the code
