+++ src/components/HomeComponenet.css
@@ -34,7 +34,9 @@
     display: flex;
     flex-direction: row;
     align-items: center;
+    justify-content: space-between;
     /* height: 80vh; */
+    width: 90%;
 }
 
 .home__results {
