+++ src/components/HomeComponent.js
@@ -1,3 +1,5 @@
+/* eslint-disable no-unused-vars */
+/* eslint-disable react-hooks/exhaustive-deps */
 import React, { useState, useEffect } from 'react';
 import './HomeComponenet.css';
 import Header from './Header';
@@ -8,15 +10,16 @@ import CircularProgress from '@mui/material/CircularProgress';
 import Box from '@mui/material/Box';
 import PlugImage from '../assets/images/plug.svg';
 import JobDetails from './JobDetails';
-import { useParams } from 'react-router-dom';
 import {getJobsData} from './api' ;
 
 import List from "../components/components/List/List"
 import { Grid } from '@mui/material';
 import Map from "../components/components/Map/Map";
 import Geocode from "react-geocode";
-// const baseURL = "http://localhost:9090/api";
-const baseURL = "https://quickbuck-api.herokuapp.com/api";
+
+
+const baseURL = "http://localhost:9090/api";
+// const baseURL = "https://quickbuck-api.herokuapp.com/api";
 
 Geocode.setApiKey("AIzaSyAsiey1SMwJZPItXSjAGKWZ87i9EvkV0-0");
 // set response language. Defaults to english.
@@ -41,7 +44,6 @@ function HomeComponent() {
 
     const { setJobs } = useAuth(); //Context 
     
-    const { jobId} = useParams();
 
     useEffect(() => {
         retrieveData();
@@ -83,23 +85,18 @@ function HomeComponent() {
         navigator.geolocation.getCurrentPosition(({ coords : {latitude, longitude} }) => {
         setLong(longitude);
         setLati(latitude);
-        setCoordinates({lat : latitude, lng : longitude});
+            setCoordinates({lat : latitude, lng : longitude});
         })
     }, []);
 
     useEffect(() => {
     console.log(coordinates,bounds);
-    ////
-  // function getCounty(latti,lngi) {
-  //   console.log("Hi");
-  //   console.log(latti,lngi);
+   
     let city, state, country;
     let countyName;
-    // let county = "Kings County"
     Geocode.fromLatLng( lati, long) 
     .then(response => {
-        // const address = response.results[0].formatted_address;
-        // let city, state, country, county;
+
         for (let i = 0; i < response.results[0].address_components.length; i++) {
           for (let j = 0; j < response.results[0].address_components[i].types.length; j++) {
             switch (response.results[0].address_components[i].types[j]) {
@@ -111,10 +108,8 @@ function HomeComponent() {
                 break;
               case "administrative_area_level_2":
                 countyName = response.results[0].address_components[i].long_name;
-                // setCounty(ccounty);
                 console.log(countyName);
                 setCounty(countyName);
-                // county = "Queens County";
                 break;
               case "country":
                 country = response.results[0].address_components[i].long_name;
@@ -125,19 +120,14 @@ function HomeComponent() {
           }
         }
     });
-  //   console.log("city");
     console.log("again"+county);
-  //   return county;
-    
-  //   // alert("Hellooo");
-  //   // return "Queens County";
-  // }
-    ////
+ 
     getJobsData(county)
       .then((data) => {
         console.log(data);
         setJobs(data);
       })
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [coordinates, bounds]);
 
    
@@ -151,13 +141,7 @@ function HomeComponent() {
                     <input className="search__bar" type="text" onChange={(event) =>handleSearch(event)} placeholder="Long Island, NY" />
                 </div>
 
-                {/* <div>
-                    <select className="search__select">
-                        <option>Category</option>
-                        <option>Technology</option>
-                    </select>
-                </div> */}
-
+              
                 <div>
                     <select className="search__select">
                         <option value="">Filter</option>
@@ -167,72 +151,58 @@ function HomeComponent() {
                 </div>
             </div>
 
-           <div className="home__body">  
-                {isLoading ? (
-                    <Box sx={{ display: 'flex' }}>
-                        <CircularProgress color="primary" className="loading" /> 
-                    </Box> 
-                ): (
-                    <div className="home__results">
-                        {allData.length > 0 && !isLoading ? 
-                            <div  className="home__results">
-                                {filteredData.map((job,index) => (
-                                    <Job 
-                                        key = {job.id}
-                                        jobId = {job.jobId}
-                                        name= {job.name}
-                                        location={job.joblocation}
-                                        price={job.price}
-                                        descp ={job.description}
-                                        author={job.author}
-                                        status={job.jobstatus}
-                                        tags = {job.tags}
-                                        county_name = {job.county_name}
-                                        website = {job.website}
-                                        rating = {job.rating}
-                                    />
-                                )).reverse()}
-                            </div>
-                            :
-                            <div className="error__div">
-                                <img className="error-img" src={PlugImage} alt="Not Found!" />
-                                {/* <div>
-                                   <h2>OOPS! </h2>
-                                   <p>We can't seem to find what you're looking for.</p>
-                                </div>
-                                <p>Error code: 404</p> */}
-
-                            </div>
-                        }
-
-                        
+            {/* Error warpping */}
+            {allData.length < 1 ?(
+                <h3>No data found</h3>
+            ):(
+                <div className="home__body">
+                    <div  className="home__results">
+                        {filteredData.map((job,index) => (
+                            <Job 
+                                key = {job.id}
+                                jobId = {job.jobId}
+                                name= {job.name}
+                                location={job.joblocation}
+                                price={job.price}
+                                descp ={job.description}
+                                author={job.author}
+                                status={job.jobstatus}
+                                tags = {job.tags}
+                                county_name = {job.county_name}
+                                website = {job.website}
+                                rating = {job.rating}
+                            />
+                        )).reverse()}
                     </div>
-                )}
 
-                <List />
-                <div>
-                    {/* <JobDetails jobId={jobId}/> */}
-
-                    <Grid container spacing={3} style = {{ width: '100%' }}> 
-                        <Grid item xs={12} md={6}  >
-                            {/* console.log("hello"); */}
-                            <List jobs = {allData} />
-                        </Grid>
-                        <Grid item xs={12} md={6}  >
+                    <div>
                         <Map
-                            
                             setCoordinates= {setCoordinates}
                             setBounds = {setBounds}
                             coordinates = {coordinates}
                             jobs = {allData}
                         />
-                    </Grid>
-                </Grid>
+                    </div>
+                
 
+                        {/* <Grid container spacing={3} style = {{ width: '100%' }}> 
+                            <Grid item xs={3} md={6}  >
+                                <List jobs = {allData} />
+                            </Grid>
+                            <Grid item xs={12} md={6}  >
+                                <Map
+                                    
+                                    setCoordinates= {setCoordinates}
+                                    setBounds = {setBounds}
+                                    coordinates = {coordinates}
+                                    jobs = {allData}
+                                />
+                            </Grid>
+                        </Grid> */}
                 </div>
-           </div>
-
+            )}
         </div>
+
     )
 }
