+++ src/components/api/index.js
@@ -3,7 +3,8 @@ import Geocode from "react-geocode";
 
 
 // const URL = 'https://testapijobslocation.herokuapp.com/jobs';
-const URL = 'https://quickbuck-api.herokuapp.com/api/jobs/findbycounty';
+const URL = "http://localhost:9090/api/jobs/findbycounty";
+// const URL = 'https://quickbuck-api.herokuapp.com/api/jobs/findbycounty';
